% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SweepFeatures.R
\name{SweepFeatures}
\alias{SweepFeatures}
\title{SweepFeatures}
\usage{
SweepFeatures(path, chromosome, start, finish)
}
\arguments{
\item{path}{Datasets}

\item{chromosome}{Chromosome Number}

\item{start}{Start Number}

\item{finish}{Finish Number}
}
\value{
\itemize{
  \item df: Results
}
}
\description{
SweepFeatures
}
\examples{
\donttest{
library("SweepDiscovery")
Path<-file.path(system.file("exdata/ExampleVCF.vcf.gz", package = "SweepDiscovery"))
pred<-SweepFeatures(path=Path,"1",20253,1976067)
}
}
\references{
\itemize{
\item Pavlidis, P., Alachiotis, N. A survey of methods and tools to detect recent and strong positive selection. J of Biol Res-Thessaloniki 24, 7 (2017). https://doi.org/10.1186/s40709-017-0064-0
}
}
