% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SweepPrediction.R
\name{SweepPrediction}
\alias{SweepPrediction}
\title{SweepPrediction}
\usage{
SweepPrediction(Traindata = NULL, Newdata)
}
\arguments{
\item{Traindata}{Dataset for training}

\item{Newdata}{New data for prediction}
}
\value{
\itemize{
  \item Prediction: Results
}
}
\description{
SweepPrediction
}
\examples{
\donttest{
library("SweepDiscovery")
data <- system.file("extdata", "data.csv", package = "SweepDiscovery")
Data<- read.csv(data)
pred<-SweepPrediction(Traindata=NULL,Newdata=Data)
}
}
\references{
\itemize{
\item Pavlidis, P., Alachiotis, N. A survey of methods and tools to detect recent and strong positive selection. J of Biol Res-Thessaloniki 24, 7 (2017). https://doi.org/10.1186/s40709-017-0064-0
}
}
